Fixing bugs

+++ functions.h
@@ -1,8 +1,8 @@
 #include <iostream>
 using namespace std;
 
-int takeNum(int x);
-int adding(int x, int y);
-int substract(int x, int y);
-int multiply(int x, int y);
-int division(int x, int y);
+double takeNum(double x);
+double adding(double x, double y);
+double substract(double x, double y);
+double multiply(double x, double y);
+double division(double x, double y);
